Fix 728 issues from whole-library review across all backends
Whole-library deep review (correctness, security, performance, architecture)
surfaced 987 findings; 728 were confirmed via adversarial verification and are
fixed here. All buildable backends (core+Python, CPU, CUDA, ROCm, Vulkan,
OneAPI) compile clean; MPS fixes applied but unverifiable on this host.

Highlights:
- 17 criticals: Vulkan eig divergent barriers, radix-sort wgSize OOB,
  embedding_bag_bf16 heap-OOB, bf16 optimizer shaders, MPS duplicate kernels,
  ROCm hipRAND double-free + caching-allocator self-deadlock, FSDP auto-wrap
  recursive parameter sharding.
- Cross-cutting themes: stride/non-contiguous handling in reductions and
  shape ops; Float16/BFloat16-as-Float32 buffer over-reads; Vulkan packed-half
  non-atomic RMW races; BatchNorm training-mode backward (f16/bf16/f64,
  channels>=256); additive-vs-multiplicative attention masking NaNs; FFT/
  special-math vendor-vs-fallback parity; untrusted-input parsing hardening
  (ONNX, model loaders, RPC); autograd grad formulas; optimizer/distributed
  correctness.
- New GPU shaders for correct f64/bf16 paths (batchnorm grad_input, bincount,
  quantile/trapezoid) and a public Module::replace_module for fusion.
- New backend-parity tests: nested LayerNorm non-power-of-two, quantized conv2d
  grid-z overflow.

Verified by building all enabled targets (0 errors) and running affected tests:
1314 cross-backend parity (CPU/CUDA/ROCm/Vulkan/OneAPI) + unit + targeted
multidtype/precision suites, all passing.

+++ include/tenzor/autograd/ops.hpp
@@ -1534,7 +1534,7 @@ auto nanmean(const Variable& input,
 
 /** @brief masked_fill(x, mask, value): overwrite x[mask] with scalar value.
  *         Backward: grad_x = where(mask, 0, grad_y); value is non-diff scalar. */
-auto masked_fill(const Variable& input, const Tensor& mask, float value) -> Variable;
+auto masked_fill(const Variable& input, const Tensor& mask, double value) -> Variable;
 
 /** @brief masked_select(x, mask): flat gather at mask=true positions.
  *         Backward scatters grad_y back into zeros_like(x). */
